Plant Disease Prediction with transfer learning/requirements.txt
tensorflow>=2.0
pillow
opencv-python
numpy
flask
flask-cors